start seed Gauge system
- Added seedgauge struct
- updated siloSettings
    - reduced variables to fit everything in one slot
- change milestoneStem to store the untruncated version

+++ protocol/contracts/beanstalk/init/InitBipNewSilo.sol
@@ -22,12 +22,12 @@ contract InitBipNewSilo {
 
     AppStorage internal s;
 
-    uint32 constant private BEAN_SEEDS_PER_BDV = 2e6;
-    uint32 constant private BEAN_3CRV_SEEDS_PER_BDV = 4e6;
-    uint32 constant private UNRIPE_BEAN_SEEDS_PER_BDV = 0;
-    uint32 constant private UNRIPE_BEAN_3CRV_SEEDS_PER_BDV = 0;
+    uint24 constant private BEAN_SEEDS_PER_BDV = 2e6;
+    uint24 constant private BEAN_3CRV_SEEDS_PER_BDV = 4e6;
+    uint24 constant private UNRIPE_BEAN_SEEDS_PER_BDV = 0;
+    uint24 constant private UNRIPE_BEAN_3CRV_SEEDS_PER_BDV = 0;
     
-    uint32 constant private STALK_ISSUED_PER_BDV = 10000;
+    uint16 constant private STALK_ISSUED_PER_BDV = 10000;
     
     
     function init() external {
@@ -44,7 +44,7 @@ contract InitBipNewSilo {
 
         //update all silo info for current Silo-able assets
 
-        uint32 currentSeason = s.season.current;
+        uint24 currentSeason = uint24(s.season.current);
 
         s.ss[C.BEAN].stalkEarnedPerSeason = BEAN_SEEDS_PER_BDV;
         s.ss[C.BEAN].stalkIssuedPerBdv = STALK_ISSUED_PER_BDV;
